% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compositions.R
\name{.estQshr}
\alias{.estQshr}
\title{Compute shrunk proportions using the 'qsh' method described by Ionas Erb
\href{https://doi.org/10.48550/arXiv.2205.09215}{here}.}
\usage{
.estQshr(x, cols_as_features = FALSE)
}
\arguments{
\item{x}{A table of compositional data.}

\item{cols_as_features}{A boolean. Toggles whether rows or columns are samples.}
}
\value{
A relative count table shrunk towards the arithmetic mean.
}
\description{
Shrink relative count data using the arithmetic mean.
}
