% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-params.R
\name{.getBetaMean}
\alias{.getBetaMean}
\title{Calculate the mean from a beta distribution.}
\usage{
.getBetaMean(k, n)
}
\arguments{
\item{k}{An integer. How many counts were observed}

\item{n}{An integer. How many observations were made}
}
\description{
See rbeta. Called by .getBetaMeans.
}
