% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-params.R
\name{.getBetaMeans}
\alias{.getBetaMeans}
\title{Calculate by feature mean from a beta distributions}
\usage{
.getBetaMeans(count_sample, log_transformed = F)
}
\arguments{
\item{count_sample}{A vector of count data.}

\item{log_transformed}{A boolean, whether to return log-transformed values.}
}
\description{
See rbeta.
}
