% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-params.R
\name{.getCLRMeans}
\alias{.getCLRMeans}
\title{Calculate by feature mean from a CLR transformed approximation}
\usage{
.getCLRMeans(count_sample)
}
\arguments{
\item{count_sample}{A vector of count data.}
}
\description{
See rbeta.
}
