% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compositions.R
\name{.getTableVars}
\alias{.getTableVars}
\title{Estimate the approximated vars from a CLR-transformed count sample.}
\usage{
.getTableVars(count_table, CLR_transformed = TRUE, cols_as_features = FALSE)
}
\arguments{
\item{count_table}{A table of count data with rows as features and columns as
samples.}

\item{CLR_transformed}{A boolean, whether to return centered log-ratio (CLR)
values.}

\item{cols_as_features}{A boolean. Toggles whether rows or columns are
samples.}
}
\description{
See rbeta.
}
