% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compositions.R
\name{.sCLR}
\alias{.sCLR}
\title{Shrink and CLR-transform a relative count table.}
\usage{
.sCLR(count_table, cols_as_features = FALSE)
}
\arguments{
\item{count_table}{A table of count data with rows as features and columns as samples.}

\item{cols_as_features}{A boolean. Toggles whether rows or columns are samples.}
}
\description{
Compute shrunk proportions using the 'qsh' method described by Ionas Erb \href{https://doi.org/10.48550/arXiv.2205.09215}{here}.
}
