% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{g_center_by}
\alias{g_center_by}
\title{Center by the geometric mean.}
\usage{
g_center_by(x, MARGIN = MARGIN)

g_center_by(x, MARGIN = MARGIN)
}
\arguments{
\item{x}{a table}

\item{MARGIN}{see apply}
}
\description{
Simply divide each value by its corresponding geometric mean.
}
