% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-params.R
\name{getLogBetaVar}
\alias{getLogBetaVar}
\title{Calculate the log variance from a beta distribution.}
\usage{
getLogBetaVar(k, n)
}
\arguments{
\item{k}{An integer. How many counts were observerd}

\item{n}{An integer. How many obserations were made}
}
\description{
See rbeta. Called by .getBetaVars.
}
