% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{impute_and_scale}
\alias{impute_and_scale}
\title{Impute zeroes and scale to sum to 1.}
\usage{
impute_and_scale(x, type = "auto", m = 0.65)
}
\arguments{
\item{x}{a numeric table.}

\item{type}{One of c("auto", "numeric", "integer"). What type of numbers are represented}

\item{m}{Defaults to 0.65. The scaling factor to impute zeroes in the non-integer case.}
}
\description{
Scales both count data and generic compositional data to proportions
}
