% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-extend.R
\name{mock_parameters}
\alias{mock_parameters}
\title{Generate mock parameters for testing}
\usage{
mock_parameters(
  n_features = 10,
  persistence = 0.95,
  interaction_sd = 0.01,
  volatility = 0.05,
  correlation = 0.5
)
}
\arguments{
\item{n_features}{numeric. Number of features in the dataset.}

\item{persistence}{numeric. Persistence parameter for the dynamics matrix.}

\item{interaction_sd}{numeric. Standard deviation of the interaction terms in the dynamics matrix.}

\item{volatility}{numeric. Volatility parameter for the process covariance matrix.}

\item{correlation}{numeric. Correlation parameter for the process covariance matrix.}
}
\value{
A list containing the mock parameters.
}
\description{
Generate mock parameters for testing the orrey_extend function.
}
