% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-random.R
\name{sampleCLRApprox}
\alias{sampleCLRApprox}
\title{samples from the approximated probability density function of a CLR-transformed count sample.}
\usage{
sampleCLRApprox(samples, count_sample)
}
\arguments{
\item{samples}{An integer. How many samples should be taken}

\item{count_sample}{A vector of count data.}
}
\description{
samples from the approximated probability density function of a CLR-transformed count sample.
}
