% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-random.R
\name{sampleCLRdiff}
\alias{sampleCLRdiff}
\title{samples from the probability density function of the difference of two
CLR-transformed count samples.}
\usage{
sampleCLRdiff(samples, x, y)
}
\arguments{
\item{samples}{An integer. How many samples should be taken}

\item{x}{A vector of count data.}

\item{y}{A second vector of count data.}
}
\description{
samples from the probability density function of the difference of two
CLR-transformed count samples.
}
