% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-random.R
\name{sampleGeomMeam}
\alias{sampleGeomMeam}
\title{samples from the log() of the geometric mean.}
\usage{
sampleGeomMeam(samples, count_sample, log_transformed = T)
}
\arguments{
\item{samples}{An integer. How many samples should be taken}

\item{count_sample}{A vector of count data.}

\item{log_transformed}{A boolean, whether to return log-transformed values.}
}
\description{
samples from the log() of the geometric mean.
}
